1. Fixes the issue of incomplete (top row and right-most column) polygon filling in the scanline polygon filling algorithm described in Ticket #1140.
2. Removed prefix randomization from the temporary files created by stsci.skypac.skymatch: skymatch masks now have fixed fixed names.

This change was due to complains from INS users that adrodrizzle leaves a large number of temporary files that keeps growing after each run. The reason is that users turn 'clean' off because they *do* want to inspect other temporary files, but they do not want these files to grow in number.


git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci.skypac/trunk@32129 fe389314-cf27-0410-b35b-8c050e845b92